another int that should be float

+++ Aquaria/Entity.h
@@ -281,7 +281,7 @@ public:
 	float getHealthPerc();
 	void setDeathScene(bool v);
 	bool isDeathScene() const { return deathScene; }
-	void generateCollisionMask(int ovrCollideRadius=0);
+	void generateCollisionMask(float ovrCollideRadius=0);
 	DamageData lastDamage;
 	bool checkSplash(const Vector &override=Vector(0,0,0));
 	EatData eatData;
